refactoring camera actions away from main script
